fix: running from fresh install
