+++ src/package.json
@@ -35,7 +35,8 @@
         "plugin.link": "npm link && cd ../demo && npm link nativescript-webview-ext && cd ../src",
         "plugin.prepare": "npm run tsc && cd ../demo && tns plugin remove nativescript-webview-ext && tns plugin add ../src",
         "plugin.tscwatch": "npm run build:webview-bridge && npm run tsc -- -w",
-        "postinstall": "([ -f './node_modules/.bin/ts-node' ] && ts-node ./postinstall.ts) || ([ -f ./postinstall.js ] && node ./postinstall) || exit 0",
+        "prepare": "npm run build:polyfill && ts-node ./make-bridge-loader.ts",
+        "postinstall": "([ -f ./postinstall.js ] && node ./postinstall) || exit 0",
         "run.demo.android": "cd ../demo && tns run android --syncAllFiles",
         "run.demo.ios": "cd ../demo && tns run ios --syncAllFiles",
         "run.http-server": "cd ../demo/www && npx hs -p 8080 --cors",
